refactor(cymbal): split ResolverConfig into core, move resolver builder
Extract the symbol-resolution config slice into core::config::ResolverConfig
(database, object storage, frame cache, sourcemap, symbol-resolution
concurrency, internal_api_secret) and nest it on the processing Config via
`#[envconfig(nested = true)]`, so env-var names are unchanged. Move
build_symbol_resolver/build_catalog + get_aws_config + FRAME_CONTEXT_LINES
out of the processing app_context into core (core::resolver / core::config).

Symbolication constructors and their tests now take &ResolverConfig instead
of the monolithic &Config; resolution mode and the remote-resolution client
read config.resolver.*. This removes resolution mode's dependency on the
processing app_context and on processing-only config fields — the key
crate-extraction blocker.

No behavior change; cargo check + clippy green. Phase 2 of 3.
