+++ rust/cymbal/src/app_context.rs
@@ -5,15 +5,14 @@ use moka::future::{Cache, CacheBuilder};
 use rdkafka::producer::FutureProducer;
 use sqlx::{postgres::PgPoolOptions, PgPool};
 use std::{sync::Arc, time::Duration};
-use tokio::{
-    sync::{Mutex, Semaphore},
-    task::JoinHandle,
-};
+use tokio::{sync::Semaphore, task::JoinHandle};
 use tracing::info;
 use uuid::Uuid;
 
 use crate::{
-    config::{get_aws_config, init_global_state, Config},
+    config::{init_global_state, Config},
+    core::config::get_aws_config,
+    core::resolver::build_catalog,
     error::UnhandledError,
     signals::{MaybeSignalClient, SignalClient},
     stages::resolution::remote::{
@@ -21,17 +20,7 @@ use crate::{
         RemoteResolutionConfig,
     },
     symbolication::symbol::{local::LocalSymbolResolver, SymbolResolver},
-    symbolication::symbol_store::{
-        apple::AppleProvider,
-        caching::{Caching, SymbolSetCache},
-        chunk_id::ChunkIdFetcher,
-        concurrency,
-        hermesmap::HermesMapProvider,
-        proguard::ProguardProvider,
-        saving::Saving,
-        sourcemap::SourcemapProvider,
-        BlobClient, Catalog, S3Client,
-    },
+    symbolication::symbol_store::{BlobClient, Catalog, S3Client},
     teams::TeamManager,
     types::operator::TeamId,
 };
@@ -73,112 +62,12 @@ impl Drop for AppContext {
     }
 }
 
-/// Build just the symbol-resolution stack from env config: connects to
-/// Postgres, builds the S3 client, calls [`init_global_state`], and
-/// returns a fully-wired `SymbolResolver`. **Does not** start Kafka
-/// producers, Redis clients, signals, the issue cache, or the remote-
-/// resolution pool — those belong to the full cymbal pipeline and are
-/// constructed by [`AppContext::from_config`].
-///
-/// Used by resolution mode (`crate::modes::resolution`), which only needs
-/// symbol resolution; keeping this separate stops resolution-mode pods from
-/// holding Kafka/Redis connections they never use.
-pub(crate) async fn build_symbol_resolver(
-    config: &Config,
-) -> Result<Arc<dyn SymbolResolver>, UnhandledError> {
-    init_global_state(config);
-    let options = PgPoolOptions::new().max_connections(config.max_pg_connections);
-    let posthog_pool = options.connect(&config.database_url).await?;
-    let s3 = aws_sdk_s3::Client::from_conf(get_aws_config(config).await);
-    let s3_client: Arc<dyn BlobClient> = Arc::new(S3Client::new(s3));
-    s3_client.ping_bucket(&config.object_storage_bucket).await?;
-    let catalog = build_catalog(config, s3_client, posthog_pool.clone());
-    Ok(Arc::new(LocalSymbolResolver::new(
-        config,
-        catalog,
-        posthog_pool,
-    )))
-}
-
-/// Build the symbol-store [`Catalog`] from already-constructed S3 and PG
-/// handles. Internal helper shared by [`AppContext::new`] and
-/// [`build_symbol_resolver`] so the provider wiring doesn't drift.
-fn build_catalog(
-    config: &Config,
-    s3_client: Arc<dyn BlobClient>,
-    posthog_pool: PgPool,
-) -> Arc<Catalog> {
-    let ss_cache = Arc::new(Mutex::new(SymbolSetCache::new(
-        config.symbol_store_cache_max_bytes,
-    )));
-
-    let smp = SourcemapProvider::new(config).with_chunk_id_rescue(
-        posthog_pool.clone(),
-        s3_client.clone(),
-        config.object_storage_bucket.clone(),
-    );
-    let smp_chunk = ChunkIdFetcher::new(
-        smp,
-        s3_client.clone(),
-        posthog_pool.clone(),
-        config.object_storage_bucket.clone(),
-    );
-    let smp_saving = Saving::new(
-        smp_chunk,
-        posthog_pool.clone(),
-        s3_client.clone(),
-        config.object_storage_bucket.clone(),
-        config.ss_prefix.clone(),
-    );
-    let smp_caching = Caching::new(smp_saving, ss_cache.clone());
-    // We want to fetch each sourcemap from the outside world exactly once,
-    // and if it isn't in the cache, load/parse it from s3 exactly once too.
-    // Limiting the per symbol set reference concurrency to 1 ensures this.
-    let smp_atmostonce = concurrency::AtMostOne::new(smp_caching);
-
-    let hmp_chunk = ChunkIdFetcher::new(
-        HermesMapProvider {},
-        s3_client.clone(),
-        posthog_pool.clone(),
-        config.object_storage_bucket.clone(),
-    );
-    // Skip the saving layer for HermesMapProvider, since it'll never fetch
-    // something from the outside world.
-    let hmp_caching = Caching::new(hmp_chunk, ss_cache.clone());
-    let hmp_atmostonce = concurrency::AtMostOne::new(hmp_caching);
-
-    let pgp_chunk = ChunkIdFetcher::new(
-        ProguardProvider {},
-        s3_client.clone(),
-        posthog_pool.clone(),
-        config.object_storage_bucket.clone(),
-    );
-    let pgp_caching = Caching::new(pgp_chunk, ss_cache.clone());
-    let pgp_atmostonce = concurrency::AtMostOne::new(pgp_caching);
-
-    let apple_chunk = ChunkIdFetcher::new(
-        AppleProvider {},
-        s3_client.clone(),
-        posthog_pool.clone(),
-        config.object_storage_bucket.clone(),
-    );
-    let apple_caching = Caching::new(apple_chunk, ss_cache);
-    let apple_atmostonce = concurrency::AtMostOne::new(apple_caching);
-
-    Arc::new(Catalog::new(
-        smp_atmostonce,
-        hmp_atmostonce,
-        pgp_atmostonce,
-        apple_atmostonce,
-    ))
-}
-
 impl AppContext {
     pub async fn from_config(config: &Config) -> Result<Self, UnhandledError> {
-        let options = PgPoolOptions::new().max_connections(config.max_pg_connections);
-        let posthog_pool = options.connect(&config.database_url).await?;
+        let options = PgPoolOptions::new().max_connections(config.resolver.max_pg_connections);
+        let posthog_pool = options.connect(&config.resolver.database_url).await?;
 
-        let s3_client = aws_sdk_s3::Client::from_conf(get_aws_config(config).await);
+        let s3_client = aws_sdk_s3::Client::from_conf(get_aws_config(&config.resolver).await);
         let s3_client = S3Client::new(s3_client);
         let s3_client = Arc::new(s3_client);
 
@@ -238,9 +127,11 @@ impl AppContext {
             _ => immediate_producer.clone(),
         };
 
-        s3_client.ping_bucket(&config.object_storage_bucket).await?;
+        s3_client
+            .ping_bucket(&config.resolver.object_storage_bucket)
+            .await?;
 
-        let catalog = build_catalog(config, s3_client, posthog_pool.clone());
+        let catalog = build_catalog(&config.resolver, s3_client, posthog_pool.clone());
 
         info!("AppContext initialized");
 
@@ -257,12 +148,12 @@ impl AppContext {
         };
 
         let symbol_resolver = Arc::new(LocalSymbolResolver::new(
-            config,
+            &config.resolver,
             catalog.clone(),
             posthog_pool.clone(),
         ));
         let symbol_resolution_limiter =
-            Arc::new(Semaphore::new(config.symbol_resolution_concurrency.max(1)));
+            Arc::new(Semaphore::new(config.resolver.symbol_resolution_concurrency.max(1)));
         let process_request_limiter =
             Arc::new(Semaphore::new(config.process_max_in_flight_requests.max(1)));
