+++ rust/cymbal/src/modes/resolution/app_context.rs
@@ -2,8 +2,8 @@ use std::sync::Arc;
 
 use tokio::sync::Semaphore;
 
-use crate::app_context::build_symbol_resolver;
 use crate::config::Config as CymbalConfig;
+use crate::core::resolver::build_symbol_resolver;
 use crate::error::UnhandledError;
 use crate::symbolication::symbol::SymbolResolver;
 
@@ -25,10 +25,11 @@ pub struct ResolutionAppContext {
 
 impl ResolutionAppContext {
     pub async fn from_config(config: &CymbalConfig) -> Result<Self, UnhandledError> {
-        let symbol_resolver = build_symbol_resolver(config).await?;
+        let symbol_resolver = build_symbol_resolver(&config.resolver).await?;
         // Symbol-resolution concurrency is the shared knob on the parent config.
-        let symbol_resolution_limiter =
-            Arc::new(Semaphore::new(config.symbol_resolution_concurrency.max(1)));
+        let symbol_resolution_limiter = Arc::new(Semaphore::new(
+            config.resolver.symbol_resolution_concurrency.max(1),
+        ));
         let load_monitor = LoadMonitor::new(config.resolution.max_item_concurrency.max(1) as u32);
         let service_instance_id = config
             .resolution
