+++ rust/cymbal/tests/resolve.rs
@@ -111,11 +111,11 @@ async fn end_to_end_resolver_test() {
     }
 
     let mut config = Config::init_with_defaults().unwrap();
-    config.allow_internal_ips = true; // We're hitting localhost for the tests
+    config.resolver.allow_internal_ips = true; // We're hitting localhost for the tests
 
-    let sourcemap = SourcemapProvider::new(&config);
+    let sourcemap = SourcemapProvider::new(&config.resolver);
     let cache = Arc::new(Mutex::new(SymbolSetCache::new(
-        config.symbol_store_cache_max_bytes,
+        config.resolver.symbol_store_cache_max_bytes,
     )));
 
     let wrapped = NoOpChunkIdFetcher { inner: sourcemap };
