+++ rust/cymbal/tests/utils.rs
@@ -48,7 +48,7 @@ pub(crate) async fn get_response_with_config<T: for<'de> Deserialize<'de>>(
     configure: impl FnOnce(&mut Config),
 ) -> (StatusCode, T) {
     let mut config = Config::init_with_defaults().unwrap();
-    config.object_storage_bucket = storage_bucket.clone();
+    config.resolver.object_storage_bucket = storage_bucket.clone();
     configure(&mut config);
 
     let issue_buckets_redis_client = Arc::new(MockRedisClient::new());
@@ -83,7 +83,7 @@ pub(crate) async fn get_raw_response(
     s3_client: Arc<MockS3Client>,
 ) -> (StatusCode, String) {
     let mut config = Config::init_with_defaults().unwrap();
-    config.object_storage_bucket = storage_bucket.clone();
+    config.resolver.object_storage_bucket = storage_bucket.clone();
 
     let issue_buckets_redis_client = Arc::new(MockRedisClient::new());
